requirements.txt
black==22.3.0
eth-ape==0.8.10
vyper==0.3.7
eth-typing==3.5.2